dont use cargo workspaces
apparently there is a bug/misfeature which prevents us from setting
the panic strategy for a specific package in the workspace, which
is broken because then running `cargo check` from the workspace root
fails, because it doesn't take into account the panic strategy of the
inner packages.

instead just use seperate crates which are in a shared folder.