improve the way we represent memory regions
the current representation did not allow representing a memory region which extends all the way to the end of the
address space, because it required providing a valid end address.

but, if the address range extends to the end of the 4GB virtual address space, then its end address can't fit in a
32 bit integer.

so, we now instead use a start address and a size to represent this.

this is still problematic, since we can not create a region of size 4GB, but for now this is good enough.

in the future, we might want to represent these regions by using an inclusive end address, which will provide maximal
flexibility.